refactor: Move disable all but core mods button to repair window
