+++ src-vue/src/views/DeveloperView.vue
@@ -42,10 +42,6 @@
 
             <h3>Repair:</h3>
 
-            <el-button type="primary" @click="disableAllModsButCore">
-                Disable all but core mods
-            </el-button>
-
             <el-button type="primary" @click="getInstalledMods">
                 Get installed mods
             </el-button>
@@ -111,28 +107,6 @@ export default defineComponent({
         },
         async launchGameWithoutChecks() {
             this.$store.commit('launchGame', true);
-        },
-        async disableAllModsButCore() {
-            let game_install = {
-                game_path: this.$store.state.game_path,
-                install_type: this.$store.state.install_type
-            } as GameInstall;
-            await invoke("disable_all_but_core", { gameInstall: game_install }).then((message) => {
-                ElNotification({
-                    title: 'Success',
-                    message: "Disabled all mods but core",
-                    type: 'success',
-                    position: 'bottom-right'
-                });
-            })
-                .catch((error) => {
-                    ElNotification({
-                        title: 'Error',
-                        message: error,
-                        type: 'error',
-                        position: 'bottom-right'
-                    });
-                });
         },
         async getInstalledMods() {
             let game_install = {
